requirements.txt
google-generativeai>=0.8.5
python-dotenv>=1.0.0
ffmpeg-python>=0.2.0
pyaudio>=0.2.14
SpeechRecognition>=3.14.2 